feat(jobs): refactor types and enum location
