+++ lib/jobs/services/email-job.service.ts
@@ -15,7 +15,7 @@
 import 'server-only';
 
 import { jobDispatcher } from '../job-dispatcher.service';
-import { JOB_TYPES } from '../types';
+import { JOB_TYPES } from '../../types/jobs';
 import type { SendEmailJobPayload } from '../schemas/send-email-job.schema';
 import type { BaseJobMetadata } from '../schemas/base-job.schema';
