--- lib/jobs/types/index.ts
@@ -1,3 +1,0 @@
-export * from './job-type.enum';
-export * from './job-config.type';
-export * from './job-registry';
