+++ lib/types/jobs/enums/job-type.enum.ts
